Creation/modification Evaluation Rubrique

+++ src/main/java/Demo/DAO/EvaluationDAO.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface EvaluationDAO extends JpaRepository<Evaluation, Integer>{
+public interface EvaluationDAO extends JpaRepository<Evaluation, Long>{
 
 }
